tests: Fix popen test on AIX/PASE
Like Solaris, popen for a non-existent file does not fail, but
causes sh to produce an error message instead. Unfortunately,
the error messages don't match with Solaris and the order of the
output for the resource and the error from sh are swapped, so
we can't just use the sunos variant of this test.